Pull request review suggestions applied #87
