+++ src/main/groovy/org/mockito/release/internal/gradle/AutoVersioningPlugin.java
@@ -4,15 +4,7 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.gradle.api.tasks.Exec;
 import org.mockito.release.internal.gradle.util.TaskMaker;
-
-import org.mockito.release.gradle.ReleaseConfiguration;
-import org.mockito.release.gradle.SecureExecTask;
-import org.mockito.release.internal.gradle.util.GitUtil;
-
-import static org.mockito.release.internal.gradle.configuration.DeferredConfiguration.deferredConfiguration;
-import static org.mockito.release.internal.gradle.configuration.LazyConfiguration.lazyConfiguration;
 
 /**
  * Plugin uses bumping version in version.properties file done by VersioningPlugin
@@ -28,23 +20,23 @@ import static org.mockito.release.internal.gradle.configuration.LazyConfiguratio
  * Adds following tasks:
  *
  * <ul>
- *     <li>bumpVersionAndPush</li>
+ *     <li>performVersionBump</li>
  * </ul>
  */
 public class AutoVersioningPlugin implements Plugin<Project> {
 
-    static final String BUMP_VERSION_AND_PUSH_TASK = "bumpVersionAndPush";
+    static final String PERFORM_VERSION_BUMP = "performVersionBump";
 
     public void apply(final Project project) {
-        project.getPlugins().apply(GitPushPlugin.class);
+        project.getPlugins().apply(GitPlugin.class);
         project.getPlugins().apply(VersioningPlugin.class);
 
 
-        TaskMaker.task(project, BUMP_VERSION_AND_PUSH_TASK, new Action<Task>() {
+        TaskMaker.task(project, PERFORM_VERSION_BUMP, new Action<Task>() {
             public void execute(Task t) {
                 t.setDescription("Increments version number, commits and pushes changes to Git repository");
                 t.dependsOn(VersioningPlugin.BUMP_VERSION_FILE_TASK);
-                t.dependsOn(GitPushPlugin.PERFORM_GIT_PUSH_TASK);
+                t.dependsOn(GitPlugin.PERFORM_GIT_PUSH_TASK);
             }
         });
     }
