--- src/main/resources/META-INF/gradle-plugins/org.mockito.mockito-release-tools.git-push.properties
@@ -1,1 +1,0 @@
-implementation-class=org.mockito.release.internal.gradle.GitPushPlugin
