--- src/test/groovy/org/mockito/release/internal/gradle/GitPushPluginTest.groovy
@@ -1,14 +1,0 @@
-package org.mockito.release.internal.gradle
-
-import org.gradle.testfixtures.ProjectBuilder
-import spock.lang.Specification
-
-class GitPushPluginTest extends Specification {
-
-    def project = new ProjectBuilder().build()
-
-    def "applies"() {
-        expect:
-        project.plugins.apply("org.mockito.mockito-release-tools.git-push")
-    }
-}
